Create a new project named CubeJump
Create a new project named CubeJump, now one game a week is begin!
Some special thing is the camera add image effect, I find Unity image
effect named DepthOfField.cs is not work on Android Phone, so I wirte a
new scripts for myself, welcome to submit for me!
I think this game have some more to need to change for lets it to be
better! For example: the better score, and I most wanted to do is make
some model for player cube!hahaha!